Fix rules-fidelity and persistence issues from full-codebase review
Rules fixes: correct three occupations' Credit Rating ranges (Police Officer,
Hacker, Tribal Member), add the missing Zealot occupation, model the book's
"either X or Y" skill-point formulas as a real choice instead of a hardcoded
characteristic, stop awarding experience ticks on bonus-die rolls, award
+2D6 Sanity when a skill crosses 90% in the development phase, forbid
bonus/penalty dice on Sanity rolls and clamp Sanity to its computed max, and
give Wealth decimal precision so the Penniless tier can pay $0.50.

Code-quality fixes: never delete a saved character on a JSON parse
failure (a future schema change would otherwise wipe every save on load),
refresh characteristic-derived skill bases (Dodge, Language (Own)) whenever
derived stats recompute so they can't go stale, let weapons link live to a
sheet skill instead of duplicating its value, and clamp age-related
characteristic deductions to what the target stats can actually absorb.

+++ CthulhuSheets/Models/Wealth.cs
@@ -2,7 +2,10 @@ namespace CthulhuSheets.Models;
 
 public class Wealth
 {
-    public int? SpendingLevel { get; set; }
-    public int? Cash { get; set; }
+    // Decimal because the 1920s Penniless band is $0.50 cash / $0.50 spending
+    // level (Table II). Existing saves hold whole-number JSON values, which
+    // deserialize into decimal without loss.
+    public decimal? SpendingLevel { get; set; }
+    public decimal? Cash { get; set; }
     public List<string> Assets { get; set; } = [];
 }
